ENH: Add custom graphics support to Tobii
ENH: Add custom graphics support to Tobii
RF: Tobii calibration now extends BaseCalibrationProcedure
